# public/api.yaml
openapi: 3.0.0
info:
  title: "Jawen"
  description: |
   API pour le site Jawen shop
  version: "1.0.0"
  termsOfService: "http://127.0.0.1:3000/"
  contact:
    name: "Support API"
    url: "http://127.0.0.1:3000/"
    email: "support@example.com"
  license:
    name: "MIT License"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://127.0.0.1:3000/api/v1"
    description: "Staging server"
paths:
...
  /dettes/{id}/paiements:
    get:
      summary: Get payments related to a debt
      operationId: getPayments
      tags:
        - Dettes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of payments related to the debt
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Paiement'
  /dettes/{id}/articles:
    get:
      summary: Get articles related to a debt
      operationId: getArticles
      tags:
        - Dettes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of articles related to the debt
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: integer
        libelle:
          type: string
        description:
          type: string
        prix:
          type: number
          format: float
        qteStock:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Paiement:
      type: object
      properties:
        id:
          type: integer
        montant:
          type: number
          format: float
        date:
          type: string
          format: date-time

        Article:
          type: object
          properties:
            id:
              type: integer
            libelle:
              type: string
...